fix: Trivial splitting functionality into files
